mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "mcp>=1.6.0",
    "openai>=1.72.0",
    "python-dotenv>=1.1.0",
]
